requirements.txt
anthropic==0.96.0
ccxt==4.5.49
eth-account==0.13.7
hyperliquid-python-sdk==0.22.0
numpy==2.4.4
pandas==3.0.2
python-dotenv==1.2.2
requests==2.33.1
aiohttp==3.13.5
